Move addMinter tests into appropriate block
