+++ {{cookiecutter.project_name}}/tests/test_token.py
@@ -1,7 +1,7 @@
 import ape
 import pytest
 
-# Standard test comes from the interpretation of EIP-20 
+# Standard test comes from the interpretation of EIP-20
 ZERO_ADDRESS = "0x0000000000000000000000000000000000000000"
 
 
@@ -9,7 +9,7 @@ def test_initial_state(token, owner):
     """
     Test inital state of the contract.
     """
-    # Check the token meta matches the deployment 
+    # Check the token meta matches the deployment
     # token.method_name() has access to all the methods in the smart contract.
     assert token.name() == "{{cookiecutter.token_name}}"
     assert token.symbol() == "{{cookiecutter.token_symbol}}"
@@ -21,10 +21,10 @@ def test_initial_state(token, owner):
     # Check intial balance of tokens
 {%- if cookiecutter.premint == 'y' %}
     assert token.totalSupply() == {{cookiecutter.premint_amount}}
-    assert token.balanceOf(owner) == {{cookiecutter.premint_amount}} 
+    assert token.balanceOf(owner) == {{cookiecutter.premint_amount}}
 {%- else %}
     assert token.totalSupply() == 1000
-    assert token.balanceOf(owner) == 1000 
+    assert token.balanceOf(owner) == 1000
 {%- endif %}
 
 
@@ -41,7 +41,7 @@ def test_transfer(token, owner, receiver):
     assert owner_balance == 1000
 {%- endif %}
 
-    receiver_balance = token.balanceOf(receiver) 
+    receiver_balance = token.balanceOf(receiver)
     assert receiver_balance == 0
 
     # token.method_name() has access to all the methods in the smart contract.
@@ -55,7 +55,7 @@ def test_transfer(token, owner, receiver):
     assert logs[0].receiver == receiver
     assert logs[0].amount == 100
 
-    receiver_balance = token.balanceOf(receiver) 
+    receiver_balance = token.balanceOf(receiver)
     assert receiver_balance == 100
 
     owner_balance = token.balanceOf(owner)
@@ -66,13 +66,13 @@ def test_transfer(token, owner, receiver):
 {%- endif %}
 
     # Expected insufficient funds failure
-    # ape.reverts: Reverts the current call using a given snapshot ID. 
+    # ape.reverts: Reverts the current call using a given snapshot ID.
     # Allows developers to go back to a previous state.
     # https://docs.apeworx.io/ape/stable/methoddocs/api.html?highlight=revert
     with ape.reverts():
         token.transfer(owner, 200, sender=receiver)
-    
-    # NOTE: Transfers of 0 values MUST be treated as normal transfers 
+
+    # NOTE: Transfers of 0 values MUST be treated as normal transfers
     # and trigger a Transfer event.
     tx = token.transfer(owner, 0, sender=owner)
 
@@ -92,7 +92,7 @@ def test_transfer_from(token, owner, accounts):
     assert owner_balance == 1000
 {%- endif %}
 
-    receiver_balance = token.balanceOf(receiver) 
+    receiver_balance = token.balanceOf(receiver)
     assert receiver_balance == 0
 
     # Spender with no approve permission cannot send tokens on someone behalf
@@ -107,7 +107,7 @@ def test_transfer_from(token, owner, accounts):
     assert logs[0].owner == owner
     assert logs[0].spender == spender
     assert logs[0].amount == 300
-    
+
     assert token.allowance(owner, spender) == 300
 
     # With auth use the allowance to send to receiver via spender(operator)
@@ -118,14 +118,14 @@ def test_transfer_from(token, owner, accounts):
     assert logs[0].sender == owner
     assert logs[0].receiver == receiver
     assert logs[0].amount == 200
-    
+
     assert token.allowance(owner, spender) == 100
 
     # Cannot exceed authorized allowance
     with ape.reverts():
         token.transferFrom(owner, receiver, 200, sender=spender)
-    
-    token.transferFrom(owner, receiver, 100, sender=spender) 
+
+    token.transferFrom(owner, receiver, 100, sender=spender)
     assert token.balanceOf(spender) == 0
     assert token.balanceOf(receiver) == 300
 {%- if cookiecutter.premint == 'y' %}
@@ -149,11 +149,11 @@ def test_approve(token, owner, receiver):
     assert logs[0].owner == owner
     assert logs[0].spender == spender
     assert logs[0].amount == 300
-    
+
     assert token.allowance(owner, spender) == 300
 
     # Set auth balance to 0 and check attacks vectors
-    # though the contract itself shouldn’t enforce it, 
+    # though the contract itself shouldn’t enforce it,
     # to allow backwards compatibility
     tx = token.approve(spender, 0, sender=owner)
 
@@ -162,7 +162,7 @@ def test_approve(token, owner, receiver):
     assert logs[0].owner == owner
     assert logs[0].spender == spender
     assert logs[0].amount == 0
-    
+
     assert token.allowance(owner, spender) == 0
 {%- if cookiecutter.mintable == 'y' %}
 
@@ -174,7 +174,7 @@ def test_mint(token, owner, receiver):
     totalSupply = token.totalSupply()
     assert totalSupply == 1000
 
-    receiver_balance = token.balanceOf(receiver) 
+    receiver_balance = token.balanceOf(receiver)
     assert receiver_balance == 0
 
     tx = token.mint(receiver, 420, sender=owner)
@@ -185,11 +185,33 @@ def test_mint(token, owner, receiver):
     assert logs[0].receiver == receiver.address
     assert logs[0].amount == 420
 
-    receiver_balance = token.balanceOf(receiver) 
+    receiver_balance = token.balanceOf(receiver)
     assert receiver_balance == 420
 
     totalSupply = token.totalSupply()
     assert totalSupply == 1420
+
+
+def test_add_minter(token, owner, accounts):
+    """
+    Test adding new minter.
+    Must trigger MinterAdded Event.
+    Must return true when checking if target isMinter
+    """
+    target = accounts[1]
+    token.addMinter(target, sender=owner)
+    assert token.isMinter(target) is True
+
+
+def test_add_minter_targeting_zero_address(token, owner):
+    """
+    Test adding new minter targeting ZERO_ADDRESS
+    Must trigger a ContractLogicError (ape.exceptions.ContractLogicError)
+    """
+    target = ZERO_ADDRESS
+    with pytest.raises(ape.exceptions.ContractLogicError) as exc_info:
+        token.addMinter(target, sender=owner)
+    assert exc_info.value.args[0] == "Cannot add zero address as minter."
 {%- endif %}
 {%- if cookiecutter.burnable == 'y' %}
 
@@ -201,7 +223,7 @@ def test_burn(token, owner):
     totalSupply = token.totalSupply()
     assert totalSupply == 1000
 
-    owner_balance = token.balanceOf(owner) 
+    owner_balance = token.balanceOf(owner)
     assert owner_balance == 1000
 
     tx = token.burn(420, sender=owner)
@@ -211,14 +233,14 @@ def test_burn(token, owner):
     assert logs[0].sender == owner
     assert logs[0].amount == 420
 
-    owner_balance = token.balanceOf(owner) 
+    owner_balance = token.balanceOf(owner)
     assert owner_balance == 580
 
     totalSupply = token.totalSupply()
     assert totalSupply == 580
 {%- endif %}
 {%- if cookiecutter.permitable == 'y' %}
-    
+
 
 @pytest.mark.skip(reason="gas estimation reverts")
 def test_permit(chain, token, owner, receiver, Permit):
@@ -231,7 +253,7 @@ def test_permit(chain, token, owner, receiver, Permit):
     assert token.allowance(owner, receiver) == 0
     permit = Permit(owner.address, receiver.address, amount, nonce, deadline)
     signature = owner.sign_message(permit.signable_message).encode_rsv()
-    
+
     with ape.reverts():
         token.permit(receiver, receiver, amount, deadline, signature, sender=receiver)
     with ape.reverts():
@@ -240,7 +262,7 @@ def test_permit(chain, token, owner, receiver, Permit):
         token.permit(owner, receiver, amount + 1, deadline, signature, sender=receiver)
     with ape.reverts():
         token.permit(owner, receiver, amount, deadline + 1, signature, sender=receiver)
-    
+
     token.permit(owner, receiver, amount, deadline, signature, sender=receiver)
 
     assert token.allowance(owner, receiver) == 100
